refactor test_exported scipt
